requirements.txt
youtube-transcript-api==0.6.2
streamlit
google-generativeai
python-dotenv
pathlib
langchain
langchain-community
langchain_openai
openai
langchain_chroma 
langchain_core 
biopython 
fuzzywuzzy 
python-Levenshtein
pandas
aiohttp
scikit-learn
-e .
